pyproject.toml
[tool.poetry]
name = "debtwarnsys"
version = "0.1.0"
description = "TCC project - Django version"
authors = ["Dickson R <>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.5"
mysql-connector-python = "^8.3.0"
python-dotenv = "^1.0.1"
pymysql = "^1.1.0"
cryptography = "^42.0.2"
commitizen = "^3.14.1"
django-countries = "^7.6.1"
django-localflavor = "^4.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
autoflake = "^2.3.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.1"
pytest-django = "^4.8.0"
pytest-factoryboy = "^2.7.0"
faker = "^25.3.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
version_files = [
    "VERSION",
    "setup.py",
    "pyproject.toml:version"
]
bump_message = "bump: version $current_version → $new_version"
update_changelog_on_bump = true
annotated_tag = true